refactor: 🚚 fichier Door dans le dossier "statics"
